remove duplicate entries in database
but still preserve the concept that duplicate entries is due to the stats not changing
from one snapshot to the next by reintroducing one duplicate entry of the previously
established timestamp prior to insert the changed stat.

ie. i dont want so many datapoints to be removed that it might lead to misleading graphs

suppose player x has the following snapshots

10 kills on day 1
12 kills on day 2
12 kills on day 3 (no change)
12 kills on day 4 (no change)
16 kills on day 5

after removing dups, it will look like:

10 kills on day 1
12 kills on day 2
16 kills on day 5

but instead i want it to look like:

10 kills on day 1
12 kills on day 2
12 kills on day 4
16 kills on day 5

that way you get a flat line from day 2 to 4, indicating no change between that time.