so actually there was no bug. put in a bunch more debug prints to trace every step and convince myself that it works.
this is very promising... out of our 50 entries, it was able to find ~42 duplicates,
leaving us with just 8 lines of actual real data. this is a significant reduction of lines we
have to parse through to do calculations which will eventually lead us to faster leaderboard
graphing.

we're ready to move to next step. disable DEBUG messages, remove the "=" when printing out to
output.txt (was put there for debugging). finally to run it on the entire database.

oh yea forgot to do a final dump of all the last bits of what's in the hash before we exit out
otherwise we'll lose the last point of data cuz it's still sitting in the hash waiting for
an entry to come and bump it out into the file.